api/requirements.txt
# Requirements.txt MINIMAL pour API Itadaki - Python 3.11 compatible
# Seulement les dépendances utilisées dans le code

# MACHINE LEARNING & DATA
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scikit-learn>=1.3.0,<1.6.0
Pillow>=10.0.0,<11.0.0
tensorflow==2.16.2

# API FRAMEWORK
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
